Refactor TranslationDTO and add TranslationSenses
The TranslationDTO class has been refactored to include a list of TranslationSenses.
The helper functions for JSON encoding/decoding of a list of TranslationDTOs have been removed.
A new class, TranslationSenses, has been added to represent the sense, roman, and word of a translation.
Both classes now have fromJson and toJson methods for serialization.